requirements.txt
numpy==1.24.3
opencv-python==4.8.0.76
librosa==0.10.1
scipy==1.10.1
tensorflow==2.13.0
matplotlib==3.7.2
Pillow==10.0.0
smbus2==0.4.3
RPi.GPIO==0.7.1
RPLCD==1.3.1
PyAudio==0.2.13
